Move classes to their own file and make them internal
