+++ features/mpaccount/src/main/java/com/jpp/mpaccount/account/UserAccountParam.kt
@@ -1,0 +1,16 @@
+package com.jpp.mpaccount.account
+
+import android.content.res.Resources
+import com.jpp.mpaccount.R
+
+internal data class UserAccountParam(
+    val screenTitle: String,
+    val posterSize: Int
+) {
+    companion object {
+        fun create(resources: Resources, posterSize: Int) = UserAccountParam(
+            screenTitle = resources.getString(R.string.account_title),
+            posterSize = posterSize
+        )
+    }
+}
